% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_create.R
\name{quarto_create}
\alias{quarto_create}
\title{Create a TRCP Quarto report template}
\usage{
quarto_create(path = ".", filename = "report.qmd")
}
\arguments{
\item{path}{Folder where the report should be created.}

\item{filename}{Name of the qmd file.}
}
\description{
Copies the TRCP Quarto skeleton into the current project folder.
}
